# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillTwin AI Mentor API",
    "description": "Contract and intelligence source-of-truth for the SkillTwin personalized AI mentor platform. Exposes adaptive goals, winding journey roadmaps, evidence-based learner twin state, real-time mentor guidance, and spaced retention retrieval practice.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/journeys/{journey_id}": {
      "get": {
        "tags": [
          "Journeys"
        ],
        "summary": "Get Journey Roadmap",
        "description": "Returns the complete structured winding roadmap including nodes,\norder, phases, states (LOCKED, AVAILABLE, CURRENT, COMPLETED, SKIPPED),\nprerequisites, and calculated overall progress.",
        "operationId": "get_journey_api_v1_journeys__journey_id__get",
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/journeys/{journey_id}/nodes/{node_id}": {
      "patch": {
        "tags": [
          "Journeys"
        ],
        "summary": "Update Journey Node State",
        "description": "Updates the state of a specific node (e.g. COMPLETED, CURRENT).\nTriggers prerequisite cascade unlocking downstream and recalculates overall journey progress.",
        "operationId": "update_node_state_api_v1_journeys__journey_id__nodes__node_id__patch",
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          },
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneyNodeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/journeys/{journey_id}/remediation": {
      "post": {
        "tags": [
          "Journeys"
        ],
        "summary": "Insert Remediation Node",
        "description": "Dynamically inserts a remediation node before a target node when a knowledge gap is identified.",
        "operationId": "insert_remediation_api_v1_journeys__journey_id__remediation_post",
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemediationInsertRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyNodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mentor/today": {
      "get": {
        "tags": [
          "Mentor"
        ],
        "summary": "Get Today's Recommended Best Action",
        "description": "SkillTwin North Star: Returns what the learner should do right now,\nthe pedagogical reasoning ('Why this action'), daily checklist, and mentor briefing.",
        "operationId": "get_mentor_today_api_v1_mentor_today_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorTodayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mentor/message": {
      "post": {
        "tags": [
          "Mentor"
        ],
        "summary": "Send Message to AI Mentor",
        "description": "Send an empathetic, goal-aligned query to the persistent AI mentor.\nReceives contextual guidance and actionable next step recommendations.",
        "operationId": "send_mentor_message_api_v1_mentor_message_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MentorMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/maintenance": {
      "get": {
        "tags": [
          "Knowledge Maintenance"
        ],
        "summary": "Get Personalized Knowledge Maintenance Posture",
        "description": "Evaluates every concept in the learner's ecosystem against mastery, retention decay,\ngoal relevance, prerequisites, known misconceptions, recent evidence, and journey position.\n\nAssigns each concept one of five strategic maintenance actions:\n- FIX: Active misconceptions or remediation nodes requiring immediate cognitive repair\n- REVISE: Mastered or in-progress concepts suffering retention decay or due for review\n- LEARN_NEXT: Unlocked next concepts in the learner's journey\n- KEEP: Strong mastery and fresh retention requiring no immediate action\n- DEPRIORITIZE: Irrelevant to active goals or deeply locked behind prerequisites",
        "operationId": "get_knowledge_maintenance_api_v1_knowledge_maintenance_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeMaintenanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/notes/generate": {
      "post": {
        "tags": [
          "Knowledge Maintenance"
        ],
        "summary": "Generate Personalized Study Notes",
        "description": "Synthesizes custom study notes tailored directly to the learner's state using:\n- Active goal context\n- Target concept definition\n- Semantic RAG source chunks retrieved from uploaded resources\n- Learner twin cognitive state (mastery, retention)\n- Specific diagnosed misconceptions and previous explanation evaluations\n\nCached efficiently by user cognitive state to reduce redundant LLM calls.",
        "operationId": "generate_personalized_notes_api_v1_knowledge_notes_generate_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateNotesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalizedStudyNotesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionRecommendationSchema": {
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "estimated_minutes": {
            "type": "integer",
            "title": "Estimated Minutes"
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id"
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 1
          },
          "quick_action_label": {
            "type": "string",
            "title": "Quick Action Label",
            "default": "Start Now"
          }
        },
        "type": "object",
        "required": [
          "action_type",
          "title",
          "reason",
          "estimated_minutes"
        ],
        "title": "ActionRecommendationSchema"
      },
...
      "GenerateNotesRequest": {
        "properties": {
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "force_regenerate": {
            "type": "boolean",
            "title": "Force Regenerate",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "concept_id"
        ],
        "title": "GenerateNotesRequest"
      },
      "GenerationStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PROCESSING",
          "READY",
          "FAILED"
        ],
        "title": "GenerationStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JourneyEdgeResponse": {
        "properties": {
          "source_node_id": {
            "type": "string",
            "title": "Source Node Id"
          },
          "target_node_id": {
            "type": "string",
            "title": "Target Node Id"
          },
          "edge_type": {
            "type": "string",
            "title": "Edge Type",
            "default": "prerequisite"
          }
        },
        "type": "object",
        "required": [
          "source_node_id",
          "target_node_id"
        ],
        "title": "JourneyEdgeResponse"
      },
      "JourneyNodeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle",
            "default": ""
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "state": {
            "$ref": "#/components/schemas/NodeState"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          },
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "estimated_minutes": {
            "type": "integer",
            "title": "Estimated Minutes",
            "default": 20
          },
          "prerequisites": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prerequisites",
            "default": []
          },
          "is_remediation": {
            "type": "boolean",
            "title": "Is Remediation",
            "default": false
          },
          "order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order"
          },
          "position_x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position X",
            "default": 0.5
          },
          "position_y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Y",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "phase",
          "state",
          "concept_id"
        ],
        "title": "JourneyNodeResponse"
      },
      "JourneyNodeUpdate": {
        "properties": {
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NodeState"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          }
        },
        "type": "object",
        "title": "JourneyNodeUpdate"
      },
      "JourneyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "goal_id": {
            "type": "string",
            "title": "Goal Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/JourneyNodeResponse"
            },
            "type": "array",
            "title": "Nodes",
            "default": []
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/JourneyEdgeResponse"
            },
            "type": "array",
            "title": "Edges",
            "default": []
          },
          "status": {
            "$ref": "#/components/schemas/JourneyStatus",
            "default": "ACTIVE"
          },
          "generation_status": {
            "$ref": "#/components/schemas/GenerationStatus",
            "default": "READY"
          },
          "generation_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "goal_id",
          "title"
        ],
        "title": "JourneyResponse"
      },
      "JourneyStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "COMPLETED",
          "PAUSED"
        ],
        "title": "JourneyStatus"
      },
      "KnowledgeMaintenanceResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "total_concepts": {
            "type": "integer",
            "title": "Total Concepts"
          },
          "breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Breakdown"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConceptMaintenanceItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "generated_at",
          "total_concepts",
          "breakdown",
          "items"
        ],
        "title": "KnowledgeMaintenanceResponse"
      },
...
      "MentorMessageRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "example": "Why did you recommend recursion today?"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MentorMessageRequest"
      },
      "MentorMessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "$ref": "#/components/schemas/MentorRole",
            "default": "mentor"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "suggested_actions": {
            "items": {
              "$ref": "#/components/schemas/ActionRecommendationSchema"
            },
            "type": "array",
            "title": "Suggested Actions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content",
          "created_at"
        ],
        "title": "MentorMessageResponse"
      },
...
      "MentorTodayResponse": {
        "properties": {
          "greeting": {
            "type": "string",
            "title": "Greeting"
          },
          "mentor_note": {
            "type": "string",
            "title": "Mentor Note"
          },
          "current_goal_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Goal Title"
          },
          "current_goal_progress": {
            "type": "number",
            "title": "Current Goal Progress",
            "default": 0.0
          },
          "recommended_action": {
            "$ref": "#/components/schemas/ActionRecommendationSchema"
          },
          "alternative_actions": {
            "items": {
              "$ref": "#/components/schemas/ActionRecommendationSchema"
            },
            "type": "array",
            "title": "Alternative Actions"
          },
          "candidate_rankings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Candidate Rankings"
          },
          "deadline_pressure": {
            "type": "string",
            "title": "Deadline Pressure",
            "default": "normal"
          },
          "pipeline_steps_executed": {
            "type": "integer",
            "title": "Pipeline Steps Executed",
            "default": 14
          },
          "daily_checklist": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Daily Checklist"
          }
        },
        "type": "object",
        "required": [
          "greeting",
          "mentor_note",
          "recommended_action"
        ],
        "title": "MentorTodayResponse"
      },
...
      "NodeState": {
        "type": "string",
        "enum": [
          "LOCKED",
          "AVAILABLE",
          "CURRENT",
          "COMPLETED",
          "SKIPPED"
        ],
        "title": "NodeState"
      },
      "PersonalizedStudyNotesResponse": {
        "properties": {
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "key_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Points"
          },
          "your_weakness": {
            "type": "string",
            "title": "Your Weakness"
          },
          "remember_this": {
            "type": "string",
            "title": "Remember This"
          },
          "next_action": {
            "type": "string",
            "title": "Next Action"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "title",
          "summary",
          "key_points",
          "your_weakness",
          "remember_this",
          "next_action",
          "generated_at"
        ],
        "title": "PersonalizedStudyNotesResponse"
      },
...
      "RemediationInsertRequest": {
        "properties": {
          "target_node_id": {
            "type": "string",
            "title": "Target Node Id"
          },
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle"
          },
          "estimated_minutes": {
            "type": "integer",
            "title": "Estimated Minutes",
            "default": 15
          }
        },
        "type": "object",
        "required": [
          "target_node_id",
          "concept_id",
          "title",
          "subtitle"
        ],
        "title": "RemediationInsertRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}